refactor(runtime): isolate terminal ledger repair
